Add support for barrel imports suggestions
