+++ .changeset/kind-dancers-do.md
@@ -1,0 +1,5 @@
+---
+"@effect/language-service": minor
+---
+
+Add support for barrel imports suggestions
