+++ README.md
@@ -84,8 +84,10 @@ Few options can be provided alongside the initialization of the Language Service
         },
         "quickinfo": true, // controls quickinfo over Effect (default: true)
         "completions": true, // controls Effect completions (default: true)
+        "goto": true, // controls Effect goto references (default: true)
         "allowedDuplicatedPackages": [], // list of package names that has effect in peer dependencies and are allowed to be duplicated (default: [])
-        "namespaceImportPackages": [] // list of package names that should be preferred as imported with namespace imports e.g. ["effect"] (default: [])
+        "barrelImportPackages": [], // package names that should be preferred as imported from the top level barrel file (default: [])
+        "namespaceImportPackages": [] // package names that should be preferred as imported with namespace imports e.g. ["effect"] (default: [])
       }
     ]
   }
